feat: Define Request and Response Structs
This commit adds the definition of the `OAIRequest`, `OAIChoices`, and `OAIResponse` structs that we will use to interact with the OpenAI API. We use the `serde_derive` crate to derive the serialization and deserialization traits for these structs. These structs will allow us to send requests to the OpenAI API and parse the responses that we receive. This commit does not change any existing functionality.