Add LRU caching to OpenAI struct
This commit adds an LRU caching layer to the `OpenAI` struct, which is responsible for making API calls to the OpenAI API. The cache is implemented using the `lru_cache` crate and is used to store the responses to previous API calls, so that the same response can be returned without making another API call if the same request is made again.

The cache has a maximum size of 100 entries and is used to store responses for requests made with the same prompt and max_tokens values.

This caching layer should improve the performance of the `generate_code` method by reducing the number of API calls made and speeding up subsequent calls that use the same prompt and max_tokens values.