feat: Add request timeout to OpenAI API client
This commit adds request timeout functionality to the `generate_code` method of the `OpenAI` struct to prevent hanging indefinitely in case the API endpoint fails to respond. The timeout is set to 10 seconds by default, but can be adjusted via the `timeout` method of the `reqwest::ClientBuilder` struct.

A new HTTP client is created with the `timeout` option set, and is used to make the API request. If the request takes longer than the specified timeout duration, an error is returned.

Note that in case of a timeout, the error message returned is a generic "API call failed" message, which could be improved in the future if necessary.